# json/swagger_pretty.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HelpHub API",
    "description": "API Documentation for HelpHub application",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/verify-2fa": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "verify2fa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register user",
        "description": "Registers a new user in the system",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or validation errors",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login user",
        "description": "Authenticates a user and returns a JWT token",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VerifyOtpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "description": "The verification token received from the send endpoint",
            "example": "a1b2c3d4e5f6..."
          },
          "otp": {
            "type": "string",
            "description": "The 6-digit OTP code",
            "example": 123456,
            "maxLength": 6,
            "minLength": 6
          }
        },
        "required": [
          "email",
          "otp",
          "token"
        ]
      },
      "ApiResponseAuthResponse": {
        "type": "object",
        "description": "Generic API response wrapper",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "Indicates whether the API request was successful or not",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable message providing more details about the outcome",
            "example": "Operation completed successfully"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Map of field/property names to arrays of error messages, populated during validation failures"
          },
          "data": {
            "$ref": "#/components/schemas/AuthResponse",
            "description": "The payload/data returned by the endpoint, shape depends on the exact endpoint"
          },
          "status_code": {
            "type": "string",
            "description": "Internal application-specific status code mapping to business logic outcomes",
            "enum": [
              "SUCCESS",
              "TWO_FA_REQUIRED",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "VALIDATION_ERROR",
              "INTERNAL_SERVER_ERROR",
              "BAD_REQUEST"
            ],
            "example": "SUCCESS"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "description": "Response containing authentication token",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT authentication token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "twoFactorRequired": {
            "type": "boolean"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "description": "Request object for user registration",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User's first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User's last name",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "User's email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User's password",
            "example": "SecurePass123!"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "description": "Request object for user login",
        "properties": {
          "email": {
            "type": "string",
            "description": "User's email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User's password",
            "example": "SecurePass123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Enter your JWT token",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}